third_party/WebKit/LayoutTests/printing/multicol.html

<!DOCTYPE html>
<script src="../resources/js-test.js"></script>
<p>To run this test manually, print this document.</p>
<p>There should be 2 pages in this document.</p>
<div style="page-break-before:always; -webkit-columns:12; -webkit-column-gap:0; width:12em; text-align:center; line-height:400px; orphans:1; widows:1;">
    m<br>
    u<br>
    l<br>
    t<br>
    i<br>
    c<br>
    o<br>
    l<br>
    <br>
    F<br>
    T<br>
    W<br>
</div>
<script>
    if (window.internals)
        shouldBe("internals.numberOfPages(800, 800)", "2");
</script>
