third_party/blink/web_tests/platform/generic/printing/multicol-expected.txt

PASS internals.numberOfPages(800, 800) is 2
PASS successfullyParsed is true

TEST COMPLETE


To run this test manually, print this document.

There should be 2 pages in this document.

m
u
l
t
i
c
o
l

F
T
W
